Challenge_1b/requirements.txt
PyMuPDF
pytesseract
Pillow
tqdm
